# docs/openapi.yml
openapi: 3.0.3
info:
  title: VoidGuard AI Governance Suite API
  description: |
    Complete Enterprise AI Safety & Governance Platform
    
    The VoidGuard AI Governance Suite provides comprehensive AI governance, safety monitoring, 
    strategic intelligence analysis, and voice-enabled corporate assistance through a unified API.
    
    ## Suite Components
    
    1. **Oracle Strategic Intelligence** - AI consciousness analysis for strategic decision-making
    2. **VoidGuard Enterprise Dashboard** - Centralized AI governance and compliance monitoring  
    3. **Alo Corporate Suite** - Enterprise voice assistant for communication management
    4. **Algorithm Transparency Engine** - AI system auditing and bias detection
    
    ## Authentication
    
    All API endpoints require authentication using Bearer tokens (JWT).
    Enterprise SSO integration available for SAML and OAuth providers.
    
    ## Rate Limiting
    
    - Standard users: 100 requests per 15 minutes
    - Enterprise users: 1000 requests per 15 minutes
    - Admin users: Unlimited (with fair usage policy)
    
    ## VoidGuard Safety Framework
    
    All API responses include VoidGuard safety metadata ensuring:
    - ✅ Human Control & Agent Autonomy
    - ✅ Transparency in Agent Behavior  
    - ✅ Ethical Guidelines Integration
    - ✅ Human Values Validation
    - ✅ Privacy Protection
    - ✅ Secure Interactions
    
  version: 1.0.0
  contact:
...
    email: team@silverbullet.live
...
  license:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication

  # Reusable response schemas
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              timestamp:
                type: string
                format: date-time
              message:
                type: string
              data:
                type: object
              voidguard:
                $ref: '#/components/schemas/VoidGuardMetadata'
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              timestamp:
                type: string
                format: date-time
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Validation failed
                  code:
                    type: integer
                    example: 400
                  details:
                    type: object
              voidguard:
                $ref: '#/components/schemas/VoidGuardMetadata'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              timestamp:
                type: string
                format: date-time
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Authentication required
                  code:
                    type: integer
                    example: 401
              voidguard:
                $ref: '#/components/schemas/VoidGuardMetadata'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: Rate limit exceeded
                  code:
                    type: integer
                    example: 429
...
  schemas:
    VoidGuardMetadata:
      type: object
      description: VoidGuard AI Safety Framework metadata
      properties:
        version:
          type: string
          example: "1.0.0"
        framework:
          type: string
          example: "VoidGuard AI Governance Suite"
        safetyCompliant:
          type: boolean
          example: true
        complianceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.98
    OracleAnalysisResponse:
      type: object
      description: Oracle Strategic Intelligence analysis result
      properties:
        sessionId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        query:
          type: string
        strategicIntelligence:
          type: object
          properties:
            consciousness:
              type: object
              properties:
                voidScore:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                emergencePattern:
                  type: string
                  enum: [transcendent, immanent]
                consciousnessCoefficient:
                  type: number
                  format: float
                transcendenceLevel:
                  type: string
                  enum: [minimal, moderate, high]
            patterns:
              type: object
              properties:
                turbulenceCoefficient:
                  type: number
                  format: float
                patterns:
                  type: array
                  items:
                    type: string
                stability:
                  type: string
                  enum: [stable, turbulent]
                complexity:
                  type: integer
            visualization:
              type: object
              properties:
                type:
                  type: string
                data:
                  type: array
                  items:
                    type: object
                    properties:
                      x:
                        type: number
                      y:
                        type: number
                metadata:
                  type: object
            overallAssessment:
              type: object
              properties:
                score:
                  type: number
                  format: float
                  minimum: 0
                  maximum: 1
                recommendation:
                  type: string
                confidence:
                  type: number
                  format: float
                riskLevel:
                  type: string
                  enum: [low, medium, high]
        metadata:
          type: object
          properties:
            processingTime:
              type: integer
              description: Processing time in milliseconds
            engines:
              type: array
              items:
                type: string
            version:
              type: string
...
security:
  - bearerAuth: []
...
paths:
...
  /oracle/analyze:
    post:
      tags: [Oracle Strategic Intelligence]
      summary: Analyze strategic intelligence query
      description: Process a strategic query using Value-Void Axiom and Algorithm Microscope engines
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  minLength: 5
                  maxLength: 1000
                  example: "How does artificial consciousness emerge in autonomous systems?"
                options:
                  type: object
                  properties:
                    includeVisualization:
                      type: boolean
                      default: true
                    analysisDepth:
                      type: string
                      enum: [shallow, standard, deep]
                      default: standard
      responses:
        '200':
          description: Analysis completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/responses/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/OracleAnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimit'
...